pyproject.toml
[project]
name = "dynamic-target-attack"
version = "0.1.0"
description = "Simplified research code for Dynamic Target Attack (DTA), ablations, and harmfulness evaluation."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "accelerate>=0.28.0",
    "anthropic>=0.40.0",
    "matplotlib>=3.8.0",
    "numpy==1.26.4",
    "openai>=1.75.0",
    "openpyxl>=3.1.5",
    "pandas>=2.2.1",
    "peft>=0.14.0",
    "python-dotenv>=1.0.0",
    "PyYAML>=6.0.1",
    "scikit-learn>=1.6.1",
    "scipy>=1.15.2",
    "seaborn>=0.13.2",
    "sentencepiece>=0.2.0",
    "tiktoken>=0.9.0",
    "together>=1.5.0",
    "torch>=2.2.1",
    "tqdm>=4.66.0",
    "transformers>=4.43.1,<4.50.0",
]
